split ot two classes
