+++ src/utils.ts
@@ -1,4 +1,5 @@
 import * as path from "path";
+import { Disposable } from "vscode";
 
 export function resourcePath(...paths: string[]): string {
   return path.join(__dirname, "..", "resources", ...paths);
@@ -8,4 +9,8 @@ export function unwrap<T>(value: T|undefined, tip = "?"): T {
   if (value === undefined)
     throw new Error("undefined " + tip);
   return value;
+}
+
+export function disposeAll(items: Disposable[]): any[] {
+  return items.map((d) => d.dispose());
 }
